@@ -1,11 +1,9 @@
 #!/usr/bin/env ruby
-# encoding: UTF-8
-$:.unshift File.join(File.dirname(__FILE__),'../lib') if ENV['DEBUG_SOAP4R']
 
 require 'getoptlong'
 require 'logger'
-require 'logger-application' unless defined?(Logger::Application)
 require 'wsdl/xmlSchema/xsd2ruby'
+
 
 class XSD2RubyApp < Logger::Application
 private
