@@ -356,11 +356,9 @@ private
   # Mapping.define_attr_accessor calls define_method with proc and it exhausts
   # much memory for each singleton Object.  just instance_eval instead of it.
   def define_xmlattr_accessor(obj, qname)
-    # untaint depends GenSupport.safemethodname
-    name = Mapping.safemethodname('xmlattr_' + qname.name).untaint
+    name = Mapping.safemethodname('xmlattr_' + qname.name)
     unless obj.respond_to?(name)
-      # untaint depends QName#dump
-      qnamedump = qname.dump.untaint
+      qnamedump = qname.dump
       obj.instance_eval <<-EOS
         def #{name}
           @__xmlattr[#{qnamedump}]
