@@ -177,13 +177,13 @@ private
     http.open_timeout = @connect_timeout if @connect_timeout
     http.read_timeout = @receive_timeout if @receive_timeout
     case url
-    when URI::HTTPS
+    when URI::HTTPS then
       if SSLEnabled
-	http.use_ssl = true
+	      http.use_ssl = true
       else
-	raise RuntimeError.new("Cannot connect to #{url} (OpenSSL is not installed.)")
+	      raise RuntimeError.new("Cannot connect to #{url} (OpenSSL is not installed.)")
       end
-    when URI::HTTP
+    when URI::HTTP then
       # OK
     else
       raise RuntimeError.new("Cannot connect to #{url} (Not HTTP.)")
