@@ -1,17 +1,13 @@
-# encoding: UTF-8
-# XSD4R - Charset handling with iconv.
+# SOAP4R - Extensions for Ruby 1.8.X compatibility
 # Copyright (C) 2000-2007  NAKAMURA, Hiroshi <nahi@ruby-lang.org>.
-
 # This program is copyrighted free software by NAKAMURA, Hiroshi.  You can
 # redistribute it and/or modify it under the same terms of Ruby's license;
 # either the dual license version in 2003, or any later version.
 
-module XSD
-
-class IconvCharset
-  def self.safe_iconv(to, from, str)
-    str.force_encoding(from).chars.map { |c| c.valid_encoding? && c || '?' }.join.encode(to)
+unless RUBY_VERSION >= "1.9.0"
+  class Hash
+    def key(value)
+      index(value)
+    end
   end
-end if RUBY_VERSION > '1.9'
-
 end
