@@ -6,7 +6,7 @@
 # redistribute it and/or modify it under the same terms of Ruby's license;
 # either the dual license version in 2003, or any later version.
 
-
+require 'soap/ruby18ext'
 require 'wsdl/parser'
 require 'wsdl/importer'
 require 'xsd/qname'
