@@ -32,8 +32,7 @@ class ServantSkeltonCreator
     result = ""
     if @modulepath
       result << "\n"
-      modulepath = @modulepath.respond_to?(:lines) ? @modulepath.lines : @modulepath # RubyJedi: compatible with Ruby 1.8.6 and above      
-      result << modulepath.collect { |ele| "module #{ele}" }.join("; ")
+      result << @modulepath.collect { |ele| "module #{ele}" }.join("; ")
       result << "\n\n"
     end
     if porttype.nil?
@@ -46,8 +45,7 @@ class ServantSkeltonCreator
     end
     if @modulepath
       result << "\n\n"
-      modulepath = @modulepath.respond_to?(:lines) ? @modulepath.lines : @modulepath # RubyJedi: compatible with Ruby 1.8.6 and above      
-      result << modulepath.collect { |ele| "end" }.join("; ")
+      result << @modulepath.collect { |ele| "end" }.join("; ")
       result << "\n"
     end
     result
