@@ -8,7 +8,6 @@
 
 
 require 'logger'
-require 'logger-application' unless defined?(Logger::Application)
 require 'xsd/qname'
 require 'wsdl/importer'
 require 'wsdl/soap/classDefCreator'
@@ -60,7 +59,7 @@ private
   end
 
   def create_file
-    @modulepath = @opt['module_path']
+    @modulepath = @opt['module_path'].split("::")
     create_classdef if @opt.key?('classdef')
     create_mapping_registry if @opt.key?('mapping_registry')
     create_servant_skelton(@opt['servant_skelton']) if @opt.key?('servant_skelton')
@@ -76,8 +75,6 @@ private
     @classdef_filename = @name + '.rb'
     check_file(@classdef_filename) or return
     write_file(@classdef_filename) do |f|
-      f.puts "# encoding: UTF-8"
-      f.puts "# Generated by wsdl2ruby\n"
       f << WSDL::SOAP::ClassDefCreator.new(@wsdl, @name_creator, @modulepath).dump
     end
   end
@@ -87,9 +84,7 @@ private
     @mr_filename = @name + 'MappingRegistry.rb'
     check_file(@mr_filename) or return
     write_file(@mr_filename) do |f|
-      f.puts "# encoding: UTF-8"
-      f.puts "# Generated by wsdl2ruby\n"
-      f << "require_relative '#{@classdef_filename}'\n" if @classdef_filename
+      f << "require '#{@classdef_filename}'\n" if @classdef_filename
       f << WSDL::SOAP::MappingRegistryCreator.new(@wsdl, @name_creator, @modulepath).dump
     end
   end
@@ -102,9 +97,7 @@ private
     check_file(@client_skelton_filename) or return
     write_file(@client_skelton_filename) do |f|
       f << shbang << "\n"
-      f.puts "# encoding: UTF-8"
-      f.puts "# Generated by wsdl2ruby\n"
-      f << "require_relative '#{@driver_filename}'\n\n" if @driver_filename
+      f << "require '#{@driver_filename}'\n\n" if @driver_filename
       f << WSDL::SOAP::ClientSkeltonCreator.new(@wsdl, @name_creator, @modulepath).dump(create_name(servicename))
     end
   end
@@ -114,9 +107,7 @@ private
     @servant_skelton_filename = (porttypename || @name + 'Servant') + '.rb'
     check_file(@servant_skelton_filename) or return
     write_file(@servant_skelton_filename) do |f|
-      f.puts "# encoding: UTF-8"
-      f.puts "# Generated by wsdl2ruby\n"
-      f << "require_relative '#{@classdef_filename}'\n\n" if @classdef_filename
+      f << "require '#{@classdef_filename}'\n\n" if @classdef_filename
       f << WSDL::SOAP::ServantSkeltonCreator.new(@wsdl, @name_creator, @modulepath).dump(create_name(porttypename))
     end
   end
@@ -128,10 +119,8 @@ private
     check_file(@cgi_stubFilename) or return
     write_file(@cgi_stubFilename) do |f|
       f << shbang << "\n"
-      f.puts "# encoding: UTF-8"
-      f.puts "# Generated by wsdl2ruby\n"
-      f << "require_relative '#{@servant_skelton_filename}'\n" if @servant_skelton_filename
-      f << "require_relative '#{@mr_filename}'\n" if @mr_filename
+      f << "require '#{@servant_skelton_filename}'\n" if @servant_skelton_filename
+      f << "require '#{@mr_filename}'\n" if @mr_filename
       f << WSDL::SOAP::CGIStubCreator.new(@wsdl, @name_creator, @modulepath).dump(create_name(servicename))
     end
   end
@@ -143,10 +132,8 @@ private
     check_file(@standalone_server_stub_filename) or return
     write_file(@standalone_server_stub_filename) do |f|
       f << shbang << "\n"
-      f.puts "# encoding: UTF-8"
-      f.puts "# Generated by wsdl2ruby\n"
-      f << "require_relative '#{@servant_skelton_filename}'\n" if @servant_skelton_filename
-      f << "require_relative '#{@mr_filename}'\n" if @mr_filename
+      f << "require '#{@servant_skelton_filename}'\n" if @servant_skelton_filename
+      f << "require '#{@mr_filename}'\n" if @mr_filename
       f << WSDL::SOAP::StandaloneServerStubCreator.new(@wsdl, @name_creator, @modulepath).dump(create_name(servicename))
     end
   end
@@ -158,10 +145,8 @@ private
     check_file(@servlet_stub_filename) or return
     write_file(@servlet_stub_filename) do |f|
       f << shbang << "\n"
-      f.puts "# encoding: UTF-8"
-      f.puts "# Generated by wsdl2ruby\n"
-      f << "require_relative '#{@servant_skelton_filename}'\n" if @servant_skelton_filename
-      f << "require_relative '#{@mr_filename}'\n" if @mr_filename
+      f << "require '#{@servant_skelton_filename}'\n" if @servant_skelton_filename
+      f << "require '#{@mr_filename}'\n" if @mr_filename
       f << WSDL::SOAP::ServletStubCreator.new(@wsdl, @name_creator, @modulepath).dump(create_name(servicename))
     end
   end
@@ -173,10 +158,8 @@ private
     creator.drivername_postfix = drivername_postfix
     check_file(@driver_filename) or return
     write_file(@driver_filename) do |f|
-      f.puts "# encoding: UTF-8"
-      f.puts "# Generated by wsdl2ruby\n"
-      f << "require_relative '#{@classdef_filename}'\n" if @classdef_filename
-      f << "require_relative '#{@mr_filename}'\n" if @mr_filename
+      f << "require '#{@classdef_filename}'\n" if @classdef_filename
+      f << "require '#{@mr_filename}'\n" if @mr_filename
       f << creator.dump(create_name(porttypename))
     end
   end
